services/requirements.txt
kafka-python
psycopg2
numpy
faker
